requirements.txt
httplib2==0.9.2
python-instagram==1.3.2
simplejson==3.8.1
six==1.10.0
wheel==0.24.0
dayonelib==0.2
httplib2==0.9.2
python-instagram==1.3.2
requests==2.9.1
simplejson==3.8.1
six==1.10.0
wheel==0.24.0
